requirements.txt
mcp[cli]>=1.9.0
pydantic>=2.0.0
uvicorn
httpx
pyjwt[crypto]
python-dotenv
